+++ genai/README.md
@@ -3,17 +3,11 @@
 ## Getting Started
 Start the api:
 ```sh
-fastapi dev src
+fastapi dev src/genai/app.py
 ```
 
 Re-generate api types in case of changes to openapi.yml with the following:
 ```sh
 rm -rf api # clean up
 openapi-generator-cli generate -g python-fastapi -i ../openapi.yml
-```
-
-Try out streaming!
-```sh
-fastapi dev src
-curl  "http://localhost:8000/stream?prompt=heythere" --no-buffer
 ```
